+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v1.2.12
-* Message : Added `read`/`write` for `double`.
+* Message
+    * Added `read`/`write` for `double`.
+    * Removed `read`/`write` for `uhalf`.
 * Buff : Buff array extension now uses `init_actor_default` and runs only once per actor.
 
 ### v1.2.11