+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.2.12
 * Fixed Cognition artifact incorrectly spawning with custom actors.
+* Object : Added `allow_cognant`.
 * Message
     * Added `read`/`write` for `double`.
     * Removed `read`/`write` for `uhalf`.